pyproject.toml
[project]
name = "telephone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.5.3",
    "openai>=1.102.0",
    "piper-tts>=1.3.0",
    "pyaudio>=0.2.14",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "soundfile>=0.13.1",
    "speechrecognition[faster-whisper]>=3.14.3",
    "ty>=0.0.1a19",
]

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff]
line-length = 120